+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tooltip was not updating value when date changes [OEMC-238](https://vizzuality.atlassian.net/browse/OEMC-238)
 - Better performance for opacity slider [OEMC-235](https://vizzuality.atlassian.net/browse/OEMC-235)
 - Geolocation position on page loading [OEMC-236](https://vizzuality.atlassian.net/browse/OEMC-236)
+- Loading layer from URL
 
 ## v1.0.0-alpha.4